+++ src/components/player-body-show.ts
@@ -173,9 +173,7 @@ class PlayerBodyShow extends PlayerBodyBase {
           ${this.alertError ? html`<ha-alert alert-type="error" dismissable @alert-dismissed-clicked=${this.alertErrorClear}>${this.alertError}</ha-alert>` : ""}
           ${(() => {
             if (this.player.attributes.sp_item_type == 'podcast') {
-              return (html`
-                    ${actionEpisodeSummary}
-                  `)
+              return (html`${actionEpisodeSummary}`)
             } else {
               return (html``)
             }